requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pydantic>=2.11.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.28.0 